Try to load dlls in directory until success
Loading of dlls is now little more robust.
The loader goes through whole list of dlls
until it success to load one.
It's for case there are more than one library in directory.
Until now when there were e.g. both debug and relese dlls,
application tried to load the first and when it was the wrong one
it didn't load the plugin.